Enable auto-reloading in watch mode when translations change
The loaders already call addDependency with the directory as the
argument, but the Webpack documentation states that it should be a
filename. So by passing each detected file to addDependency
individually, we gain automatic reloading for the most common case of an
existing translation file changing.

Fixes https://github.com/kirschbaum-development/laravel-translations-loader/issues/21
Fixes https://github.com/kirschbaum-development/laravel-translations-loader/issues/6